+++ source/modulo_components/test/cpp/test_lifecycle_component.cpp
@@ -26,7 +26,7 @@ protected:
   std::shared_ptr<LifecycleComponentPublicInterface> component_;
 };
 
-TEST_F(LifecycleComponentTest, AddOutput) {
+TEST_F(LifecycleComponentTest, AddRemoveOutput) {
   std::shared_ptr<State> data = make_shared_state(CartesianState::Random("test"));
   component_->add_output("test", data);
   auto outputs_iterator = component_->outputs_.find("test");
@@ -39,5 +39,8 @@ TEST_F(LifecycleComponentTest, AddOutput) {
   component_->add_output("test", new_data);
   EXPECT_EQ(component_->outputs_.at("test")->get_message_pair()->get_type(),
             modulo_core::communication::MessageType::ENCODED_STATE);
+
+  component_->remove_output("test_13");
+  EXPECT_TRUE(component_->outputs_.find("test_13") == component_->outputs_.end());
 }
 } // namespace modulo_components