+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Release Versions:
 - Correctly list rclcpp_components as build dependency (#137)
 - Add option to declare signals (#136)
 - Correct default value for user callback on add_input (#138)
+- Add option to remove signals (#139)
 
 ## 2.0.0
 ### August 05, 2022